refactor: movido calendario a /compartidos
